Ensure that size of scalar (1) is of integer type
mpi4py 4.0.0 seems to expect that counts in vector buffer specifications are of
integer type, while mpi4py 3.1.6 did not complain about counts of float type.

+++ elphmod/bravais.py
@@ -996,7 +996,7 @@ def resize(data, shape, angle=60, axes=(0, 1), period=None, polar=False,
         dtype=data.dtype)
 
     comm.Allgatherv(my_new_data,
-        (new_data, sizes * np.prod(data.shape[len(shape):])))
+        (new_data, sizes * np.prod(data.shape[len(shape):], dtype=int)))
 
     # restore original order of axes and return:
